chore: rename calculator loader to lower snake case
